requirements.txt
streamlit
langchain
langchain-community
langchain-openai
pypdf
python-dotenv
tiktoken
docarray
sentence-transformers
transformers
torch
